Fix table limit tests
