+++ src/service.rs
@@ -42,7 +42,7 @@ use tokio::time::Interval;
 mod ip_vote;
 mod query_info;
 //TODO: Update service tests
-//mod test;
+mod test;
 
 /// The types of requests to send to the Discv5 service.
 pub enum ServiceRequest {
